--- src/components/placeholder/Placeholder.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const Placeholder = () => {
-  return (
-    <div className='placeholder'>
-      <h1 className='placeholder-title'>Page does not exist.</h1>
-    </div>
-  );
-};
