+++ src/pages/homepage/HomePage.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from 'react';
+import React from 'react';
 import { useOutletContext } from 'react-router-dom';
 
 import { HeadNewsPreview } from '../../components/headNewsPreview/HeadNewsPreview';
@@ -6,33 +6,20 @@ import { AsideList } from '../../components/aside/AsideList';
 import { PreviewList } from '../../components/previewList/PreviewList';
 import { Menu } from '../../components/header/menu/Menu';
 
+import { useBurgerMenu } from '../../hooks/useBurgerMenu';
+
 import './homepage.css';
 
 export const HomePage = () => {
-  const { open } = useOutletContext();
+  const { handleOpen } = useOutletContext();
 
-  const overlayRef = useRef(null);
-  const layerRef = useRef(null);
-
-  useEffect(() => {
-    const overlay = overlayRef.current;
-    const layer = layerRef.current;
-    if (open) {
-      overlay.style.display = 'block';
-      layer.style.display = 'block';
-    }
-    return () => {
-      overlay.style.display = 'none';
-      layer.style.display = 'none';
-    };
-  }, [open]);
-
+  const { overlayRef, layerRef } = useBurgerMenu();
   return (
     <main className='home-page'>
       <div className='container-jsx'>
         <div ref={overlayRef} className='overlay'></div>
         <div ref={layerRef} className='burger-menu-layer'>
-          <Menu className='vertical-menu' />
+          <Menu className='vertical-menu' onClick={handleOpen} />
         </div>
         <HeadNewsPreview />
         <AsideList />
